# proto/payment.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "github.com/ianspire/amazing-payments/proto/payment.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/payment/healthcheck": {
      "get": {
        "operationId": "HealthCheck",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentHealthCheckResponse"
            }
          }
        },
        "tags": [
          "PaymentService"
        ]
      }
    },
    "/v1/customer": {
      "post": {
        "summary": "Customer Methods",
        "operationId": "CreateCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentCustomer"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentCreateCustomerRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    },
    "/v1/customer/{customerID}": {
      "get": {
        "operationId": "GetCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentCustomer"
            }
          }
        },
        "parameters": [
          {
            "name": "customerID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "stripe_customer_key",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "PaymentService"
        ]
      },
      "post": {
        "operationId": "UpdateCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentCustomer"
            }
          }
        },
        "parameters": [
          {
            "name": "customerID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentUpdateCustomerRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    }
  },
  "definitions": {
    "paymentCreateCustomerRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "stripe_charge_date": {
          "type": "string"
        }
      }
    },
    "paymentCustomer": {
      "type": "object",
      "properties": {
        "customerID": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "stripe_customer_key": {
          "type": "string"
        },
        "stripe_charge_date": {
          "type": "string"
        }
      },
      "title": "Customer method message group"
    },
    "paymentHealthCheckResponse": {
      "type": "object",
      "properties": {
        "isHealthy": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "paymentUpdateCustomerRequest": {
      "type": "object",
      "properties": {
        "customerID": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "stripe_customer_key": {
          "type": "string"
        },
        "stripe_charge_date": {
          "type": "string"
        }
      }
    }
  }
}